tracker: fix dynamic anchors corner case
When an anchor is invoked at module-level in the same module where
it is defined (possibly transitively through import time side-effect)
then the file->module mapping is not set.

While we are walking the stack to find a mathcing anchor, we can keep
track of occurrences of _find_and_load_helper to link the stack of
module names we are already maintaining, to the filepath that we are
getting in the stack trace, and update the mapping accordingly, which
allows us to correctly resolve anchors for modules in the import stack.